pyproject.toml
[tool.poetry]
package-mode = false
name = "an-chat"
version = "0.1.0"
description = ""
authors = ["Vladislav Andronov <68904879+AndrVLDZ@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.5"
python-dotenv = "^1.0.1"
Django = "^5.1"
psycopg2 = "^2.9.9"
python-slugify = "^8.0.4"

[tool.poetry.group.dev.dependencies]
ipython = "^8.27.0"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"